feat: настроил коллизию пули с пулей sok-48
